requirements.txt
PySide6==6.2.4
requests==2.26.0
python-dotenv==0.19.2